requirements.txt
requests
lxml
numpy
Pillow